+++ src/ser/mod.rs
@@ -1,3 +1,3 @@
-pub mod serialize;
+pub mod control_packet;
 pub mod serializer;
 pub use serializer::Error;
